+++ models/clienteModel.js
@@ -164,7 +164,24 @@ class ClienteModel {
         return await banco.ExecutaComandoNonQuery(sql, valores);
     }
 
+    async estaEmUso(id) {
+        let sql = `
+        select count(*) as total
+        from tb_pedido
+        where cli_id = ?
+    `;
+
+        let banco = new Database();
+        let rows = await banco.ExecutaComando(sql, [id]);
+
+        return rows[0].total > 0;
+    }
+
     async deletar(id) {
+
+        if (await this.estaEmUso(id)) {
+            return false;
+        }
 
         let sql = "delete from tb_cliente where cli_id = ?";
