+++ models/profissionalModel.js
@@ -129,9 +129,27 @@ class ProfissionalModel {
         return await banco.ExecutaComandoNonQuery(sql, valores);
     }
 
+    async estaEmUso(id) {
+        let sql = `
+        select count(*) as total
+        from tb_agendamento
+        where pro_id = ?
+    `;
+
+        let banco = new Database();
+        let rows = await banco.ExecutaComando(sql, [id]);
+
+        return rows[0].total > 0;
+    }
+
     async deletar(id) {
+        if (await this.estaEmUso(id)) {
+            return false;
+        }
+
         let sql = "delete from tb_profissional where pro_id = ?";
         let banco = new Database();
+
         return await banco.ExecutaComandoNonQuery(sql, [id]);
     }
 }
